requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.3.1
certifi==2025.7.14
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.12.0
django-otp==1.2.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
Flask==3.1.1
frozenlist==1.7.0
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
MarkupSafe==3.0.2
# MouseInfo==0.1.3  # Disabled - related to pywhatkit
msgpack==1.1.1
multidict==6.6.3
openai==1.98.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
# PyAutoGUI==0.9.54  # Disabled - related to pywhatkit
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
# PyGetWindow==0.0.9  # Disabled - related to pywhatkit
PyJWT==2.10.1
# PyMsgBox==1.0.9  # Disabled - related to pywhatkit
# pyperclip==1.9.0  # Disabled - related to pywhatkit
# PyRect==0.2.0  # Disabled - related to pywhatkit
# PyScreeze==1.0.1  # Disabled - related to pywhatkit
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-whois==0.8.0
# pytweening==1.2.0  # Disabled - related to pywhatkit
pytz==2025.2
# pywhatkit==5.4  # Disabled due to issues on Render server
redis==4.5.4
requests==2.31.0
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
twilio==8.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
yarl==1.20.1
